eta-system/requirements.txt
pydantic>=2.0
httpx>=0.27
google-genai>=1.0
supabase>=2.0
pandas>=2.0
python-dotenv>=1.0
beautifulsoup4>=4.12
streamlit>=1.30
protobuf>=5.0
